% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coRalODEs.R
\name{coralODEs.noROS}
\alias{coralODEs.noROS}
\title{Determine the rates of change of the state variables in a system without ROS}
\usage{
coralODEs.noROS(t, y, pars)
}
\description{
Can be used as func argument in solveCoral
}
