% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coRalODEs.R
\name{initEnv}
\alias{initEnv}
\title{Create functions of the time for the environment, for use with run_coral_ode}
\usage{
initEnv(times, L, N, X)
}
\arguments{
\item{times}{same as times for run_coral_ode}

\item{L}{same as L for init_env from coRal}

\item{N}{same as N for init_env from coRal}

\item{X}{same as X for init_env from coRal}
}
\value{
list of functions of t
}
\description{
Create functions of the time for the environment, for use with run_coral_ode
}
\examples{
init_env(seq(0,365,0.1), c(30,30,0), c(1e-6,e-6,0), c(1e-7,1e-7,0))
}
