% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coRalODEs.R
\name{initState}
\alias{initState}
\title{Create initial state for solveCoral with target fluxes jCP and rhoC}
\usage{
initState(pars)
}
\arguments{
\item{pars}{A list of model parameters in the same structure as defPars().}
}
\value{
A named numeric vector containing the initial flux and biomass
values.
}
\description{
Helper function, based off of run_coral() from the coRal package.
}
\examples{
initState(defPars())
}
\seealso{
\code{\link{solveCoral}}, \code{\link{defPars}}
}
