% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coRalODEs.R
\name{run_coral_ode}
\alias{run_coral_ode}
\title{Wrapper function to coRal}
\usage{
run_coral_ode(time, env, pars)
}
\arguments{
\item{time}{The time values at which output is desired.}

\item{env}{list of functions of time describing the environment, can
be obtained with initEnv}

\item{pars}{list of parameters in the structure of def_pars from coRal}
}
\description{
Has the same output structure as run_coral() from the coRal package
}
\examples{
time <- seq(0,365,0.1)
env <- initEnv(time, c(30,30,0), c(1e-6,1e-6,0), c(1e-7,1e-7,0))
pars <- coRal::def_pars()
run_coral_ode(time, env, pars)
}
\seealso{
\code{\link{initEnv}}
}
